[1.3.0] Backported r3989 (Tweaked showmemstat. Fixed updater minor error. Fixed some compiler warnings.) r3990 (Fixed #6957 (Map editor not saving locations of objects))

+++ branches/1.3.0/MTA10/mods/deathmatch/logic/CNetworkStats.cpp
@@ -150,7 +150,7 @@ void CNetworkStats::Draw ( void )
             iNumLines++;
 
     float fBackWidth = 310;
-    float fBackHeight = iNumLines * 15;
+    float fBackHeight = (float)( iNumLines * 15 );
     float fX = fResWidth - fBackWidth;
     float fY = 0.40f * fResHeight - iNumLines * 4;
